+++ src/pysight/validation_tools.py
@@ -24,47 +24,45 @@ def validate_line_input(dict_of_data: Dict, cols_in_data: List, num_of_lines: in
     if len(cols_in_data) == 0:
         raise ValueError('No columns in data.')
 
-    if 'Lines' in dict_of_data.keys():
-        # Verify that the input is not corrupt
-        max_change_pct = dict_of_data['Lines']['abs_time'][dict_of_data['Lines']['abs_time']\
-            .diff().pct_change(periods=1) > 0.05]
-        if len(max_change_pct) / dict_of_data['Lines'].shape[0] > 0.1\
-            and 'Frames' not in dict_of_data:
-            # Data is corrupted, and no frame channel can help us.
-            warnings.warn("""Line data was corrupt.
-                             Simulating line data using GUI's parameters.""")
-            line_delta = bins_bet_lines(line_freq=line_freq, binwidth=binwidth,
-                                        bidir=bidir)
-            dict_of_data['Lines'] = extrapolate_line_data(last_event=last_event_time,
-                                                          line_point=dict_of_data['Lines'].at[0, 'abs_time'],
-                                                          num_of_lines=num_of_lines,
-                                                          line_delta=line_delta,
-                                                          delay_between_frames=delay_between_frames,
-                                                          bidir=bidir, binwidth=binwidth,
-                                                          num_of_frames=num_of_frames)
-            return dict_of_data, line_delta
+    # Find the suitable case for this data regarding the line signal recorded
+    type_of_line_data = match_line_data_to_case(lines=dict_of_data['Lines'].loc[:, 'abs_time'],
+                                                num_of_lines=num_of_lines,
+                                                keys=list(dict_of_data.keys()))
 
-        elif len(max_change_pct) / dict_of_data['Lines'].shape[0] > 0.1\
-            and 'Frames' in dict_of_data:
-            # Data is corrupted, but we can rebuild lines on top of the frame channel
-            line_array = create_line_array(last_event_time=last_event_time, num_of_lines=num_of_lines,
-                                           num_of_frames=num_of_frames)
-            dict_of_data['Lines'] = pd.DataFrame(line_array, columns=['abs_time'], dtype='uint64')
-            line_delta = last_event_time / (num_of_lines * int(num_of_frames))
-            return dict_of_data, line_delta
+    if 'corrupt' == type_of_line_data:
+        # Data is corrupted, and no frame channel can help us.
+        line_delta = bins_bet_lines(line_freq=line_freq, binwidth=binwidth,
+                                    bidir=bidir)
+        dict_of_data['Lines'] = extrapolate_line_data(last_event=last_event_time,
+                                                      line_point=dict_of_data['Lines'].at[0, 'abs_time'],
+                                                      num_of_lines=num_of_lines,
+                                                      line_delta=line_delta,
+                                                      delay_between_frames=delay_between_frames,
+                                                      bidir=bidir, binwidth=binwidth,
+                                                      num_of_frames=num_of_frames)
+        return dict_of_data, line_delta
 
-        elif len(max_change_pct) / len(dict_of_data['Lines']) < 0.1:
-            # Data is valid. Check whether we need a 0-time line event
-            line_delta = dict_of_data['Lines'].loc[:, 'abs_time'].diff().median()
-            zeroth_line_delta = np.abs(dict_of_data['Lines'].loc[0, 'abs_time'] - line_delta)/line_delta
-            if zeroth_line_delta < 0.05:
-                dict_of_data['Lines'] = pd.DataFrame([[0] * len(cols_in_data)],
-                                                     columns=cols_in_data,
-                                                     dtype='uint64')\
-                    .append(dict_of_data['Lines'], ignore_index=True)
-            return dict_of_data, line_delta
 
-    else:  # create our own line array
+    elif 'rebuild' == type_of_line_data:
+        # Data is corrupted, but we can rebuild lines on top of the frame channel
+        line_array = create_line_array(last_event_time=last_event_time, num_of_lines=num_of_lines,
+                                       num_of_frames=num_of_frames)
+        dict_of_data['Lines'] = pd.DataFrame(line_array, columns=['abs_time'], dtype='uint64')
+        line_delta = last_event_time / (num_of_lines * int(num_of_frames))
+        return dict_of_data, line_delta
+
+    elif 'valid' == type_of_line_data:
+        # Data is valid. Check whether we need a 0-time line event
+        line_delta = dict_of_data['Lines'].loc[:, 'abs_time'].diff().median()
+        zeroth_line_delta = np.abs(dict_of_data['Lines'].loc[0, 'abs_time'] - line_delta)/line_delta
+        if zeroth_line_delta < 0.05:
+            dict_of_data['Lines'] = pd.DataFrame([[0] * len(cols_in_data)],
+                                                 columns=cols_in_data,
+                                                 dtype='uint64')\
+                .append(dict_of_data['Lines'], ignore_index=True)
+        return dict_of_data, line_delta
+
+    elif 'from_scratch' == type_of_line_data:  # create our own line array
         line_array = create_line_array(last_event_time=last_event_time, num_of_lines=num_of_lines,
                                        num_of_frames=num_of_frames)
         dict_of_data['Lines'] = pd.DataFrame(line_array, columns=['abs_time'], dtype='uint64')
@@ -127,6 +125,10 @@ def extrapolate_line_data(last_event: int, line_point: int=0,
     :param binwidth: Binwidth of multiscaler in seconds.
     :return: pd.DataFrame of line data
     """
+    # Verify input
+    if line_point / line_delta > num_of_lines:  # The first recorded line came very late
+        line_point = 0
+
     # Create the matrix containing the duplicate frame data
     delay_between_frames_in_bins = int(delay_between_frames / binwidth)
     time_of_frame = line_delta * num_of_lines \
@@ -135,7 +137,7 @@ def extrapolate_line_data(last_event: int, line_point: int=0,
     time_of_frame_mat = np.tile(time_of_frame * np.arange(num_of_frames), (num_of_lines, 1))
 
     # Create the matrix containing the duplicate line data
-    line_vec = np.arange(start=line_point, stop=line_delta * num_of_lines, step=line_delta,
+    line_vec = np.arange(start=line_point, stop=line_delta*num_of_lines, step=line_delta,
                          dtype=np.uint64)
     line_vec = np.r_[np.flip(np.arange(start=line_point, stop=0, step=-line_delta,
                                dtype=np.uint64)[1:], axis=0), line_vec]
@@ -336,3 +338,34 @@ def calc_last_event_time(dict_of_data: Dict, lines_per_frame: int=-1):
     else:
         return max(max_pmt1, max_pmt2)
 
+
+def match_line_data_to_case(lines: pd.Series, keys: list,
+                            num_of_lines: int=512) -> str:
+    """
+    Enumerate all possibilities of line data and choose the right option
+    :param lines: Line data
+    :param keys: Keys of `dict_of_data` dictionary
+    :param num_of_lines: Number of lines per frame
+    :return: String of the specific case. Either: 'corrupt', 'rebuild', 'valid' or 'from_scratch'
+    """
+    if 'Lines' in keys:
+        # Verify that the input is not corrupt
+        if lines.shape[0] < num_of_lines // 2:
+            warnings.warn("Line data was corrupt as there were too few lines.\n"
+                          "Simulating line data using GUI's parameters.")
+            return 'corrupt'
+
+        max_change_pct = lines[lines.diff().pct_change(periods=1) > 0.05]
+        if len(max_change_pct) / lines.shape[0] > 0.1 and 'Frames' not in keys:
+            warnings.warn("Line data was corrupt - the period didn't make sense.\n"
+                          "Simulating line data using GUI's parameters.")
+            return 'corrupt'
+
+        elif len(max_change_pct) / lines.shape[0] > 0.1 and 'Frames' in keys:
+            return 'rebuild'
+
+        elif len(max_change_pct) / lines.shape[0] < 0.1:
+            return 'valid'
+
+    else:
+        return 'from_scratch'